create class LoanCalculator.java

+++ src/main/java/LoanCalculator.java
@@ -2,21 +2,21 @@ import org.decimal4j.util.DoubleRounder;
 
 public class LoanCalculator {
 
-    public double monthlyPayment(int loanAmount, double loanTerm, double interestRate) {
+    public double monthlyPayment(int loanAmount, int loanTerm, double interestRate) {
         double monthlyPayment = calculateMonthlyPayment(loanAmount, loanTerm, interestRate);
         double result = DoubleRounder.round(monthlyPayment, 2);
         System.out.println("Payment Every Month: USD" + result);
         return result;
     }
 
-    public double totalPrincipalPaid(int loanAmount, double loanTerm, double interestRate) {
+    public double totalPrincipalPaid(int loanAmount, int loanTerm, double interestRate) {
         double totalPrincipalPaid = calculateMonthlyPayment(loanAmount, loanTerm, interestRate) * loanTerm;
         double result = DoubleRounder.round(totalPrincipalPaid, 2);
         System.out.println("Total of " + loanTerm + " Payments: USD" + result);
         return result;
     }
 
-    public double totalInterestPaid(int loanAmount, double loanTerm, double interestRate) {
+    public double totalInterestPaid(int loanAmount, int loanTerm, double interestRate) {
         double totalInterestPaid = calculateMonthlyPayment(loanAmount, loanTerm, interestRate) * loanTerm - loanAmount;
         double result = DoubleRounder.round(totalInterestPaid, 2);
         System.out.println("Total Interest: USD" + result);
@@ -24,7 +24,7 @@ public class LoanCalculator {
     }
 
 
-    private double calculateMonthlyPayment(int loanAmount, double loanTerm, double interestRate) {
+    private double calculateMonthlyPayment(int loanAmount, int loanTerm, double interestRate) {
         //Before calculating monthly payment, let’s find the monthly rate
         double monthlyInterestRate = interestRate / 100 / 12;
         //calculate the monthly payment
